backend/requirements.txt
asgiref==3.8.1
Django==5.1
django-environ==0.12.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
pytz==2025.2
drf-nested-routers===0.94.2